+++ src/scripts/profile_startup.py
@@ -1,10 +1,10 @@
 from textual.app import App, ComposeResult
-from textual_textarea import TextArea
+from textual_textarea import TextEditor
 
 
 class TextApp(App, inherit_bindings=False):
     def compose(self) -> ComposeResult:
-        self.ta = TextArea(
+        self.ta = TextEditor(
             text="class TextApp(App):",
             language="python",
             theme="monokai",