+++ src/textual_textarea/__main__.py
@@ -1,9 +1,11 @@
+from __future__ import annotations
+
 import sys
 
 from textual.app import App, ComposeResult
 from textual.widgets import Placeholder
 
-from textual_textarea import TextArea
+from textual_textarea import TextEditor
 
 
 class FocusablePlaceholder(Placeholder, can_focus=True):
@@ -12,7 +14,7 @@ class FocusablePlaceholder(Placeholder, can_focus=True):
 
 class TextApp(App, inherit_bindings=False):
     CSS = """
-    TextArea {
+    TextEditor {
         height: 1fr;
     }
     Placeholder {
@@ -26,22 +28,34 @@ class TextApp(App, inherit_bindings=False):
         except IndexError:
             language = "python"
         yield FocusablePlaceholder()
-        yield TextArea(
+        self.editor = TextEditor(
             language=language,
             theme="nord-darker",
             use_system_clipboard=True,
             id="ta",
         )
+        yield self.editor
 
     def on_mount(self) -> None:
-        ta = self.query_one("#ta", expect_type=TextArea)
-        ta.focus()
-        ta.word_completer = lambda x: [
-            (
-                "supercalifragilisticexpialadociousASDFASDFASFASDF FX",
+        self.editor.focus()
+
+        def _completer(prefix: str) -> list[tuple[str, str]]:
+            words = [
+                "satisfy",
+                "season",
+                "second",
+                "seldom",
+                "select",
+                "self",
+                "separate",
+                "set",
+                "space",
+                "super",
                 "supercalifragilisticexpialadocious",
-            )
-        ]
+            ]
+            return [(w, w) for w in words if w.startswith(prefix)]
+
+        self.editor.word_completer = _completer
 
 
 app = TextApp()